requirements.txt
flask
flask-cors
flask-compress
yt-dlp>=2024.3.10
requests
gunicorn
python-dotenv
